+++ src/index.ts
@@ -1,1 +1,0 @@
-console.log("COMPILE ALL THE THINGS!");